perf(evm): integrate precompile cache into pipeline execution stage
Move the precompile cache from engine-tree to reth-evm so both the
engine tree (live sync) and the pipeline execution stage (initial sync)
can share the same cache. Add an optional PrecompileCacheMap to
BasicBlockExecutor that wraps cacheable precompiles per block.

Amp-Thread-ID: https://ampcode.com/threads/T-019d6d17-07a8-7323-9b0f-dfd78034f6d2
