+++ crates/engine/tree/Cargo.toml
@@ -53,8 +53,6 @@ revm-primitives.workspace = true
 futures.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "macros"] }
-moka = { workspace = true, features = ["sync"] }
-
 # metrics
 metrics.workspace = true
 reth-metrics = { workspace = true, features = ["common"] }
